XTA: Moved test runners to different file.
